requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
cclib==1.7.2
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
dbstep==1.0.0
Django==4.1.3
dnspython==2.2.1
dotty-dict==1.3.1
emmet-core==0.38.9
fonttools==4.38.0
future==0.18.2
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflection==0.5.1
jsonschema==4.17.0
kiwisolver==1.4.4
latexcodec==2.0.1
llvmlite==0.39.1
Markdown==3.4.1
matplotlib==3.6.2
monty==2022.9.9
mp-api==0.29.4
mpmath==1.2.1
msgpack==1.0.4
nanoid==2.0.0
networkx==2.8.8
numba==0.56.4
numpy==1.23.4
ocelot @ git+https://github.com/caer200/ocelot_api.git@f61e9a96daa1116bb74fda56504c2dd13b40449f
packaging==21.3
palettable==3.3.0
pandas==1.5.1
periodictable==1.6.1
Pillow==9.3.0
Pint==0.20.1
pkgutil_resolve_name==1.3.10
plotly==5.11.0
PubChemPy==1.0.4
pybtex==0.24.0
pydantic==1.10.2
pymatgen==2022.11.7
pymongo==4.3.2
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-jsonschema-objects==0.4.1
pytz==2022.6
PyYAML==6.0
rdkit==2022.9.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.9.3
selfies==2.1.1
six==1.16.0
spglib==2.0.2
sqlparse==0.4.3
sympy==1.11.1
tabulate==0.9.0
tenacity==8.1.0
tqdm==4.64.1
typing_extensions==4.4.0
uncertainties==3.1.7
urllib3==1.26.12
zipp==3.10.0
